requirements.txt
fastapi
uvicorn
python-dotenv
llama_index
chromadb
